Adding support for systemd
Systemd doesn't support having a variabed for the executable name, so
I had to remove the NODE_EXE variable. Also, I've added support for
a defaults and sysconfig file, that work the same way for systemd
and statsd so that a sysadmin doesn't have to edit the RPM provided
and won't have to add os-specific conditional logic to config file
management.